use i32

+++ evaluation/src/def.rs
@@ -2,5 +2,5 @@ use chess::Board;
 
 pub trait Evaluator {
     fn name(&self) -> String;
-    fn evaluate(&mut self, board: &Board) -> i16;
+    fn evaluate(&mut self, board: &Board) -> i32;
 }
